requirements.txt
discord.py == 1.3.3
redis == 3.5.3